ms precision for settings and exponential back off
- transformed time based values in settings to float to gain ms
  resolution
- added setting for fuzzied exponential back off
