+++ .gitignore
@@ -24,3 +24,5 @@ TAGS
 /go/src/gopkg.in/tylerb/graceful.v1/
 /go/src/gopkg.in/yaml.v2/
 /go/src/source.xing.com/olympus/golympus/
+.idea
+.byebug_history
